clean up code, fix card selector bug
